Move task spec generation helpers from api file to workflow util file
