+++ src/utils/workflow-utils.ts
@@ -1,12 +1,17 @@
 import {
     DraftSpecsExtQuery_ByCatalogName,
-    generateCaptureDraftSpec,
     modifyDraftSpec,
 } from 'api/draftSpecs';
 import { DraftSpecQuery } from 'hooks/useDraftSpecs';
+import { isEmpty } from 'lodash';
 import { CallSupabaseResponse } from 'services/supabase';
 import { ResourceConfigDictionary } from 'stores/ResourceConfig/types';
 import { Schema } from 'types';
+import {
+    CaptureDef,
+    CaptureEndpoint,
+    MaterializationDef,
+} from '../../flow_deps/flow';
 
 // const mergeResourceConfigs = (
 //     queryData: DraftSpecQuery,
@@ -34,6 +39,91 @@ import { Schema } from 'types';
 
 //     return mergedResourceConfig;
 // };
+
+export const generateCaptureDraftSpec = (
+    endpoint: CaptureEndpoint,
+    resourceConfigs: ResourceConfigDictionary | null,
+    existingTaskData: DraftSpecsExtQuery_ByCatalogName | null
+): CaptureDef => {
+    const draftSpec: CaptureDef = isEmpty(existingTaskData)
+        ? {
+              bindings: [],
+              endpoint: {},
+          }
+        : existingTaskData.spec;
+
+    draftSpec.endpoint.connector = endpoint.connector;
+
+    if (resourceConfigs) {
+        Object.keys(resourceConfigs).forEach((collectionName) => {
+            const resourceConfig = resourceConfigs[collectionName].data;
+
+            if (Object.keys(resourceConfig).length > 0) {
+                const existingBindingIndex = draftSpec.bindings.findIndex(
+                    (binding: any) => binding.target === collectionName
+                );
+
+                if (existingBindingIndex > -1) {
+                    draftSpec.bindings[existingBindingIndex].resource = {
+                        ...resourceConfig,
+                    };
+                } else {
+                    draftSpec.bindings.push({
+                        target: collectionName,
+                        resource: {
+                            ...resourceConfig,
+                        },
+                    });
+                }
+            }
+        });
+    }
+
+    return draftSpec;
+};
+
+export const generateMaterializationDraftSpec = (
+    config: any,
+    image: string,
+    resources?: any,
+    existingTaskData?: DraftSpecsExtQuery_ByCatalogName | null
+) => {
+    const draftSpec: MaterializationDef = isEmpty(existingTaskData)
+        ? {
+              bindings: [],
+              endpoint: { connector: {} },
+          }
+        : existingTaskData.spec;
+
+    draftSpec.endpoint.connector = { config, image };
+
+    if (resources) {
+        Object.keys(resources).forEach((collectionName) => {
+            const resourceConfig = resources[collectionName].data;
+
+            if (Object.keys(resourceConfig).length > 0) {
+                const existingBindingIndex = draftSpec.bindings.findIndex(
+                    (binding: any) => binding.source === collectionName
+                );
+
+                if (existingBindingIndex > -1) {
+                    draftSpec.bindings[existingBindingIndex].resource = {
+                        ...resourceConfig,
+                    };
+                } else {
+                    draftSpec.bindings.push({
+                        source: collectionName,
+                        resource: {
+                            ...resourceConfig,
+                        },
+                    });
+                }
+            }
+        });
+    }
+
+    return draftSpec;
+};
 
 export interface SupabaseConfig {
     catalogName: string;
